// src/interfaces/ESError.ts
export interface ESError {
    // Autogenerated via elasticsearch npm typings
    // If the status is 429 it means that you can retry the document,
    // otherwise it's very likely a mapping error, and you should
    // fix the document before to try it again.
    status: any;
    error: any;
    operation: any;
    document: any;
}